# src/swagger_files/QuickCode.Demo.UserManagerModule.Api-Swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "QuickCode - Demo - UserManagerModule Api (1.0.0.0)",
    "termsOfService": "https://www.quickcode.net",
    "contact": {
      "name": "Üzeyir Apaydın",
      "url": "https://www.quickcode.net"
    },
    "license": {
      "name": "QuickCode.Net License",
      "url": "https://www.quickcode.net"
    },
    "version": "v1"
  },
  "paths": {
...
    "/api/user-manager-module/asp-net-roles/count": {
      "get": {
        "tags": [
          "AspNetRoles"
        ],
        "operationId": "api-user-manager-module-asp-net-roles-count-get",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/user-manager-module/asp-net-roles/{id}": {
      "get": {
        "tags": [
          "AspNetRoles"
        ],
        "operationId": "api-user-manager-module-asp-net-roles--id--get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AspNetRolesDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "AspNetRoles"
        ],
        "operationId": "api-user-manager-module-asp-net-roles--id--put",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AspNetRolesDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/AspNetRolesDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/AspNetRolesDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "AspNetRoles"
        ],
        "operationId": "api-user-manager-module-asp-net-roles--id--delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/user-manager-module/asp-net-roles/{aspNetRolesId}/asp-net-user-roles": {
      "get": {
        "tags": [
          "AspNetRoles"
        ],
        "operationId": "api-user-manager-module-asp-net-roles--aspNetRolesId--asp-net-user-roles-get",
        "parameters": [
          {
            "name": "aspNetRolesId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AspNetRolesAspNetUserRoles_RESTResponseDto"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/user-manager-module/asp-net-roles/{aspNetRolesId}/asp-net-user-roles/{aspNetUserRolesUserId}": {
      "get": {
        "tags": [
          "AspNetRoles"
        ],
        "operationId": "api-user-manager-module-asp-net-roles--aspNetRolesId--asp-net-user-roles--aspNetUserRolesUserId--get",
        "parameters": [
          {
            "name": "aspNetRolesId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "aspNetUserRolesUserId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AspNetRolesAspNetUserRoles_KEY_RESTResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/user-manager-module/asp-net-roles/{aspNetRolesId}/asp-net-role-claims": {
      "get": {
        "tags": [
          "AspNetRoles"
        ],
        "operationId": "api-user-manager-module-asp-net-roles--aspNetRolesId--asp-net-role-claims-get",
        "parameters": [
          {
            "name": "aspNetRolesId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AspNetRolesAspNetRoleClaims_RESTResponseDto"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/user-manager-module/asp-net-roles/{aspNetRolesId}/asp-net-role-claims/{aspNetRoleClaimsId}": {
      "get": {
        "tags": [
          "AspNetRoles"
        ],
        "operationId": "api-user-manager-module-asp-net-roles--aspNetRolesId--asp-net-role-claims--aspNetRoleClaimsId--get",
        "parameters": [
          {
            "name": "aspNetRolesId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "aspNetRoleClaimsId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AspNetRolesAspNetRoleClaims_KEY_RESTResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AspNetRolesAspNetRoleClaims_KEY_RESTResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "roleId": {
            "type": "string",
            "nullable": true
          },
          "claimType": {
            "type": "string",
            "nullable": true
          },
          "claimValue": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "AspNetRolesAspNetRoleClaims_RESTResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "roleId": {
            "type": "string",
            "nullable": true
          },
          "claimType": {
            "type": "string",
            "nullable": true
          },
          "claimValue": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "AspNetRolesAspNetUserRoles_KEY_RESTResponseDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "nullable": true
          },
          "roleId": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "AspNetRolesAspNetUserRoles_RESTResponseDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "nullable": true
          },
          "roleId": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "AspNetRolesDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "normalizedName": {
            "type": "string",
            "nullable": true
          },
          "concurrencyStamp": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "ApiKey": {
        "type": "apiKey",
        "description": "Please enter a valid Api Key",
        "name": "X-Api-Key",
        "in": "header"
      }
    }
  },
  "security": [
    {
      "ApiKey": [ ]
    }
  ]
}